[feat] 1:1 문의 채팅 기능 구현중

+++ src/main/java/com/hello/travelogic/inquiry/dto/InquiryChatDTO.java
@@ -11,7 +11,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @ToString
 public class InquiryChatDTO {
-    private Integer icId;
+    private Long icId;
     private Long memberCode;
     private Integer authorityCode;
     private String memberId;
